requirements.txt
langchain_openai
langchain_community
langchain_core
python-dotenv
streamlit